Fixed(?) freetype for windows machines

+++ Cubic.Freetype/FT_BBox.cs
@@ -5,8 +5,8 @@ namespace Cubic.Freetype;
 [StructLayout(LayoutKind.Sequential)]
 public struct FT_BBox
 {
-    public long XMin;
-    public long YMin;
-    public long XMax;
-    public long YMax;
+    public int XMin;
+    public int YMin;
+    public int XMax;
+    public int YMax;
 }
