+++ Cubic.Freetype/FT_Vector.cs
@@ -5,8 +5,8 @@ namespace Cubic.Freetype;
 [StructLayout(LayoutKind.Sequential)]
 public struct FT_Vector
 {
-    private long _x;
-    private long _y;
+    private int _x;
+    private int _y;
 
     public int X => (int) (_x >> 6);
     public int Y => (int) (_y >> 6);
